Break monthly reports component and logic out
